+++ client/src/components/LandingPage.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { Icon, Image } from "semantic-ui-react";
 import { Grid, Header } from "semantic-ui-react";
-import SideBarMenu from "./SideBar";
+import SideBarMenu from "./NavBarAndSideBar/SideBar";
 
 export default class landingPage extends React.Component {
   state = {
