+++ client/src/components/Opportunities/OpportunitiesList.js
@@ -11,15 +11,15 @@ import {
   Button,
   Divider
 } from "semantic-ui-react";
-import { deleteOpportunityAndConnectedSkills } from "../api/opportunities";
-import { getOpportunitiesForList, getSkillsList } from "../api/opportunities";
-import { getSkills } from "../api/skills";
-import { getCities } from "../api/cities";
+import { deleteOpportunityAndConnectedSkills } from "../../api/opportunities";
+import { getOpportunitiesForList, getSkillsList } from "../../api/opportunities";
+import { getSkills } from "../../api/skills";
+import { getCities } from "../../api/cities";
 import OpportunityTypeFilters from "./OpportunityTypeFilters";
-import { filterOpportunities } from "../utils/filterOpportunities";
+import { filterOpportunities } from "../../utils/filterOpportunities";
 import OpportunityCard from "./OpportunityCard";
 import { Link } from "react-router-dom";
-import { getLoggedInUserData } from "../utils/storage";
+import { getLoggedInUserData } from "../../utils/storage";
 
 class OpportunitiesList extends Component {
   state = {
